requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.7
gunicorn==19.9.0
MarkupSafe==1.1.1
Werkzeug==0.15.2
Click==7.0
Pillow==5.4.1
requests==2.21.0
jsonschema==3.0.1
numpy==1.16.2
pandas==0.24.2
